Allow the user final say with the CLANG_FORMAT_DIFF variable
I'd like to run my clang-format-diff with a special Python, and the
previous code would strip out my special Python and run it with whatever
'python' binary was in the path.

This allows the user final say over how to run clang-format-diff, and
brings the behaviour in line with what the documentation implies.